intercheangability tests: order fields type fixed

+++ src/MagentoAccess/Models/Services/Rest/v2x/SalesOrderRepository/RootObject.cs
@@ -699,7 +699,7 @@ namespace MagentoAccess.Models.Services.Rest.v2x.SalesOrderRepository
 		public int? base_discount_canceled { get; set; }
 		public int? base_discount_invoiced { get; set; }
 		public int? base_discount_refunded { get; set; }
-		public int? base_grand_total { get; set; }
+		public decimal? base_grand_total { get; set; }
 		public int? base_discount_tax_compensation_amount { get; set; }
 		public int? base_discount_tax_compensation_invoiced { get; set; }
 		public int? base_discount_tax_compensation_refunded { get; set; }
@@ -712,9 +712,9 @@ namespace MagentoAccess.Models.Services.Rest.v2x.SalesOrderRepository
 		public int? base_shipping_refunded { get; set; }
 		public int? base_shipping_tax_amount { get; set; }
 		public int? base_shipping_tax_refunded { get; set; }
-		public int? base_subtotal { get; set; }
+		public decimal? base_subtotal { get; set; }
 		public int? base_subtotal_canceled { get; set; }
-		public int? base_subtotal_incl_tax { get; set; }
+		public decimal? base_subtotal_incl_tax { get; set; }
 		public int? base_subtotal_invoiced { get; set; }
 		public int? base_subtotal_refunded { get; set; }
 		public int? base_tax_amount { get; set; }
@@ -722,7 +722,7 @@ namespace MagentoAccess.Models.Services.Rest.v2x.SalesOrderRepository
 		public int? base_tax_invoiced { get; set; }
 		public int? base_tax_refunded { get; set; }
 		public int? base_total_canceled { get; set; }
-		public int? base_total_due { get; set; }
+		public decimal? base_total_due { get; set; }
 		public int? base_total_invoiced { get; set; }
 		public int? base_total_invoiced_cost { get; set; }
 		public int? base_total_offline_refunded { get; set; }
@@ -763,7 +763,7 @@ namespace MagentoAccess.Models.Services.Rest.v2x.SalesOrderRepository
 		public string ext_order_id { get; set; }
 		public int? forced_shipment_with_invoice { get; set; }
 		public string global_currency_code { get; set; }
-		public int? grand_total { get; set; }
+		public decimal? grand_total { get; set; }
 		public int? discount_tax_compensation_amount { get; set; }
 		public int? discount_tax_compensation_invoiced { get; set; }
 		public int? discount_tax_compensation_refunded { get; set; }
@@ -800,9 +800,9 @@ namespace MagentoAccess.Models.Services.Rest.v2x.SalesOrderRepository
 		public string store_name { get; set; }
 		public int? store_to_base_rate { get; set; }
 		public int? store_to_order_rate { get; set; }
-		public int? subtotal { get; set; }
+		public decimal? subtotal { get; set; }
 		public int? subtotal_canceled { get; set; }
-		public int? subtotal_incl_tax { get; set; }
+		public decimal? subtotal_incl_tax { get; set; }
 		public int? subtotal_invoiced { get; set; }
 		public int? subtotal_refunded { get; set; }
 		public int? tax_amount { get; set; }
@@ -810,7 +810,7 @@ namespace MagentoAccess.Models.Services.Rest.v2x.SalesOrderRepository
 		public int? tax_invoiced { get; set; }
 		public int? tax_refunded { get; set; }
 		public int? total_canceled { get; set; }
-		public int? total_due { get; set; }
+		public decimal? total_due { get; set; }
 		public int? total_invoiced { get; set; }
 		public int? total_item_count { get; set; }
 		public int? total_offline_refunded { get; set; }
